refactor: move context utilities to dedicated `custom-context` directory
