+++ src/lib/custom-context/was-paper-reviewed-context.ts
@@ -1,0 +1,14 @@
+import { getContext, setContext } from "svelte";
+
+type WasReviewedState = { wasReviewed: boolean };
+/**
+ * Key for context storing a reactive state whether the project paper, the context is attached to was already
+ * reviewed by the user currently logged in or not.
+ */
+export const WAS_PROJECT_PAPER_ALREADY_REVIEWED_KEY = Symbol("wasProjectPaperAlreadyReviewed");
+export function setAlreadyReviewedContext(alreadyReviewed: WasReviewedState) {
+    setContext(WAS_PROJECT_PAPER_ALREADY_REVIEWED_KEY, alreadyReviewed);
+}
+export function getAlreadyReviewedContext() {
+    return getContext(WAS_PROJECT_PAPER_ALREADY_REVIEWED_KEY) as WasReviewedState;
+}
