+++ src/routes/+layout.svelte
@@ -4,7 +4,7 @@
     import { Toaster } from "svelte-sonner";
     import type { LayoutData } from "./$types";
     import { setContext, type Snippet } from "svelte";
-    import { UserContextKey, type UserContext } from "$lib/current-user/userContext";
+    import { UserContextKey, type UserContext } from "$lib/custom-context/user-context";
     import type { User } from "$lib/model/api/user";
     import { IS_IN_DEV_MODE } from "$lib/constants";
