+++ tests/integration/criteria/criterion-list-entry.test.ts
@@ -4,7 +4,7 @@ import { describe, expect, test } from "vitest";
 import { createReviewedCriterion } from "../../model-builder";
 import { Criteria, Reviews, Users } from "../../example-data";
 import { reviewMode } from "$lib/global-state/review-mode-state.svelte";
-import { SELECTED_REVIEW_CRITERIA_KEY } from "$lib/utils/custom-context";
+import { SELECTED_REVIEW_CRITERIA_KEY } from "$lib/custom-context/selected-review-criteria-context";
 import userEvent from "@testing-library/user-event";
 import { mockSelectedCriteriaContext } from "$tests/integration/test-helper";
